move H.264 NAL logic to separate header, to share with H.265
